tested without cor

+++ models/ha_sampling.py
@@ -14,7 +14,7 @@ def ha_sampling(train, test, seq_len=12, sampling_rate=2, pre_len=3, repeat=Fals
 
     result_y = []
     for i in range(len(test_x)):
-        a = np.array(test_x[i])
+        a = np.array(test_x[i], dtype='float')
 
         if repeat:
             a_mean = np.mean(a, axis=0)
@@ -42,8 +42,8 @@ def ha_sampling(train, test, seq_len=12, sampling_rate=2, pre_len=3, repeat=Fals
         test_y = temp_test_y
         result_y = temp_result_y
 
-    test_y1 = np.array(test_y)
-    result_y1 = np.array(result_y)
+    test_y1 = np.array(test_y, dtype='float')
+    result_y1 = np.array(result_y, dtype='float')
     num_nodes = test_y1.shape[-1]
     result1 = np.reshape(result_y1, [-1, num_nodes]).T
     test1 = np.reshape(test_y1, [-1, num_nodes]).T
